Trait wasmtime_cli_flags::opt::WasmtimeOptionValue

pub trait WasmtimeOptionValue: Sized {
    const VAL_HELP: &'static str;

    // Required method
    fn parse(val: Option<&str>) -> Result<Self>;
}

A helper trait for all types of options that can be parsed. This is what actually parses the =val in key=val

Required Associated Constants

const VAL_HELP: &'static str

Help text for the value to be specified.

Required Methods

fn parse(val: Option<&str>) -> Result<Self>

Parses the provided value, if given, returning an error on failure.