Struct wasmtime_cli_flags::CommonOptions

pub struct CommonOptions {

    pub config: Option<PathBuf>,
    pub disable_logging: bool,
    pub log_to_files: bool,
    pub debug_info: bool,
    pub disable_cache: bool,
    pub disable_parallel_compilation: bool,
    pub wasm_features: Option<WasmFeatures>,
    pub wasi_modules: Option<WasiModules>,
    pub jitdump: bool,
    pub vtune: bool,
    pub optimize: bool,
    pub opt_level: Option<OptLevel>,
    pub cranelift_set: Vec<(String, String)>,
    pub cranelift_enable: Vec<String>,
    pub static_memory_maximum_size: Option<u64>,
    pub static_memory_forced: bool,
    pub static_memory_guard_size: Option<u64>,
    pub dynamic_memory_guard_size: Option<u64>,
    pub enable_cranelift_debug_verifier: bool,
    pub enable_cranelift_nan_canonicalization: bool,
    pub fuel: Option<u64>,
    pub epoch_interruption: bool,
    pub disable_address_map: bool,

}

Common options for commands that translate WebAssembly modules

Fields
config: Option<PathBuf>

Use specified configuration file

disable_logging: bool

Disable logging.

log_to_files: bool

Log to per-thread log files instead of stderr.

debug_info: bool

Generate debug information

disable_cache: bool

Disable cache system

disable_parallel_compilation: bool

Disable parallel compilation

wasm_features: Option<WasmFeatures>

Enables or disables WebAssembly features

wasi_modules: Option<WasiModules>

Enables or disables WASI modules

jitdump: bool

Generate jitdump file (supported on –features=profiling build)

vtune: bool

Generate vtune (supported on –features=vtune build)

optimize: bool

Run optimization passes on translated functions, on by default

opt_level: Option<OptLevel>

Optimization level for generated functions Supported levels: 0 (none), 1, 2 (most), or s (size); default is “most”

cranelift_set: Vec<(String, String)>

Set a Cranelift setting to a given value. Use wasmtime settings to list Cranelift settings for a target.

cranelift_enable: Vec<String>

Enable a Cranelift boolean setting or preset. Use wasmtime settings to list Cranelift settings for a target.

static_memory_maximum_size: Option<u64>

Maximum size in bytes of wasm memory before it becomes dynamically relocatable instead of up-front-reserved.

static_memory_forced: bool

Force using a “static” style for all wasm memories.

static_memory_guard_size: Option<u64>

Byte size of the guard region after static memories are allocated.

dynamic_memory_guard_size: Option<u64>

Byte size of the guard region after dynamic memories are allocated.

enable_cranelift_debug_verifier: bool

Enable Cranelift’s internal debug verifier (expensive)

enable_cranelift_nan_canonicalization: bool

Enable Cranelift’s internal NaN canonicalization

fuel: Option<u64>

Enable execution fuel with N units fuel, where execution will trap after running out of fuel.

Most WebAssembly instructions consume 1 unit of fuel. Some instructions, such as nop, drop, block, and loop, consume 0 units, as any execution cost associated with them involves other instructions which do consume fuel.

epoch_interruption: bool

Executing wasm code will yield when a global epoch counter changes, allowing for async operation without blocking the executor.

disable_address_map: bool

Disables the on-by-default address map from native code to wasm code.

Implementations

impl CommonOptions

pub fn parse_from_str(s: &str) -> Result<Self>

pub fn init_logging(&self)

pub fn config(&self, target: Option<&str>) -> Result<Config>

pub fn enable_wasm_features(&self, config: &mut Config)

pub fn opt_level(&self) -> OptLevel